working

+++ app/Http/Controllers/EmployeeController.php
@@ -349,19 +349,25 @@ class EmployeeController extends Controller
             $contract->delete();
             return response()->download($pdfPath)->deleteFileAfterSend(true);
         } else {
-            return redirect()->route('/home')->back()->with('error', 'No contract found.');
+            return redirect()->route('home')->back()->with('error', 'No contract found.');
         }
     }
 
-    public function renew($employee)
+    public function renew($employeeId)
     {
-        $employee = Employee::findOrFail($employee);
-        $contract = $employee->employee_contracts();
-        DB::table('position_to_employee_contract')->where('employee_contract_id',$contract->id)->delete();
-        DB::table('holiday_saldos')->where('employee_contract_id', $contract->id)->delete();
-        $contract->delete();
-
-        return redirect()->route('/employee_add_contract');
+        $employee = Employee::findOrFail($employeeId);
+        $contract = $employee->employee_contracts()->first();
+        
+        if ($contract) {
+            DB::table('position_to_employee_contract')->where('employee_contract_id', $contract->id)->delete();
+            DB::table('holiday_saldos')->where('employee_contract_id', $contract->id)->delete();
+            $contract->delete();
+            
+            return redirect()->route('contract.index');
+        } else {
+            return redirect()->route('/ontract.index')->with('error', 'No contract found.');
+        }
+        
     }
 
     public function update(UpdateEmployeeRequest $request, Employee $employee): RedirectResponse
